requirements.txt
fastapi
uvicorn
torch
transformers
accelerate
huggingface_hub
hf_xet
pymysql
python-dotenv
cryptography